Add files via upload
Creating the E-commerce Sales Data Report Using Power BI has been an exciting personal project, showcasing the culmination of my knowledge and skills acquired while working with PowerBl. This report offers valuable insights into the sales analysis of an online goods-selling company.  The analyses and visualisation of the data provided a comprehensive understanding of the company's performance in the e-commerce market."